+++ CHANGELOG.md
@@ -24,9 +24,9 @@ All user visible changes to this project will be documented in this file. This p
         - `Create` method for `Room`, `Member`, `Endpoint`;
         - `Get` method for `Room`, `Member`, `Endpoint`;
         - `Delete` method for `Room`, `Member`, `Endpoint`.
-    - gRPC Control API callbacks:
-        - `on_join` ([#63]);
-        - `on_leave` ([#63]).
+    - gRPC Control API callbacks ([#63]):
+        - `on_join`;
+        - `on_leave`.
 - Signalling:
     - Dynamic `Peer`s creation when client connects ([#28]);
     - Auto-removing `Peer`s when `Member` disconnects ([#28]);